back/requirements2.txt
amqp
annotated-types
asgiref
attrs
billiard
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cryptography
dj-rest-auth
Django
django-allauth
django-filter
django-jazzmin
djangorestframework
djangorestframework_simplejwt
drf-spectacular
exceptiongroup
idna
inflection
jsonschema
jsonschema-specifications
kombu
packaging
pillow
prompt_toolkit
psycopg2-binary 
psycopg2
pycparser
pydantic
pydantic-settings
pydantic_core
PyJWT
python-dateutil
python-dotenv
PyYAML
referencing
requests
rest-framework-simplejwt
rpds-py
sib-api-v3-sdk
six
sqlparse
typing-inspection
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
vine
wcwidth
